requirements.txt
streamlit==1.32.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.0
duckduckgo-search==5.3.0
google-generativeai
thefuzz==0.22.1
wikipedia==1.4.0
requests==2.31.0
tensorflow
